pyproject.toml
[project]
name = "basic-browser-use"
version = "0.1.0"
description = "A very basic tool employing Browser Use."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use==0.1.40",
    "langchain-google-genai==2.1.2",
    "playwright==1.51.0",
]


[tool.setuptools]
py-modules = ["browser_tools", "preprocessor", "custom_agent", "custom_client"]

